uppdate all

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: otait-ta <otait-ta@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 10:20:38 by otait-ta          #+#    #+#             */
-/*   Updated: 2022/10/11 12:48:39 by otait-ta         ###   ########.fr       */
+/*   Updated: 2022/10/24 11:17:31 by otait-ta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	l;
